# root/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/faults": {
      "post": {
        "operationId": "FaultsController_create",
        "summary": "Create Fault",
        "description": "Logs a new equipment fault, optionally linked to an alert. Organization Admin and Technician can create faults.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFaultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fault created successfully."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "faults"
        ]
      },
      "get": {
        "operationId": "FaultsController_findAll",
        "summary": "List All Faults",
        "description": "Retrieves all fault records. Organization Admin and Technician can view faults.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of fault records returned."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "faults"
        ]
      }
    },
    "/api/faults/{id}": {
      "get": {
        "operationId": "FaultsController_findOne",
        "summary": "Get Fault by ID",
        "description": "Retrieves a single fault by UUID. Organization Admin and Technician can look up faults.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fault record returned."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "faults"
        ]
      },
      "put": {
        "operationId": "FaultsController_put",
        "summary": "Replace Fault",
        "description": "Completely replaces an existing fault record. Only Organization Admin can perform full replacements.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutFaultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "faults"
        ]
      },
      "patch": {
        "operationId": "FaultsController_update",
        "summary": "Update Fault",
        "description": "Partially updates a fault (e.g., status, reassigning Technician). Organization Admin and Technician can update.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFaultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "faults"
        ]
      },
      "delete": {
        "operationId": "FaultsController_remove",
        "summary": "Delete Fault",
        "description": "Permanently removes a fault record. Only the Organization Admin can delete faults.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "faults"
        ]
      }
    },
    "/api/faults/{id}/work-orders": {
      "get": {
        "operationId": "FaultsController_getWorkOrders",
        "summary": "Get Work Orders for Fault",
        "description": "Retrieves all work orders linked to a fault. Organization Admin and Technician can view.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of work order records returned."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "faults"
        ]
      }
    },
...
  },
  "info": {
    "title": "EnerTrack API",
    "description": "Backend API for EnerTrack — Facility Management, Sustainability & Finance Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      },
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id"
      }
    },
    "schemas": {
...
      "CreateFaultDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "alert_id": null,
          "assigned_to_id": "550e8400-0003-4000-8000-000000000003",
          "asset_name": "Chiller Unit C-12",
          "fault_type": "Mechanical",
          "severity": "moderate",
          "status": "pending"
        }
      },
      "PutFaultDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateFaultDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "alert_id": null,
          "assigned_to_id": "550e8400-0003-4000-8000-000000000003",
          "asset_name": "Chiller Unit C-12",
          "fault_type": "Mechanical",
          "severity": "moderate",
          "status": "pending"
        }
      },
...
    }
  }
}